+++ CHANGELOG.md
@@ -59,6 +59,7 @@
 - Replace the region tree with a new vue component !2483 @alex.simm
 - New vue component for blog posts !2676 !2687 @alex.simm
 - Changed forum post response to empty content #1499 !2638 @krauterisator
+- Refactor add pickup modal !2693 @chriswalg
 - Refactor edit pickup rules to vue component and moved it to storeInformationModal !2686 @chriswalg 
 
 ## Dev/Test/CI stuff